feat: repair grid display on all pages
