+++ src/components/App/App.js
@@ -3,7 +3,7 @@ import { Routes, Route } from "react-router-dom"
 import Home from '../../pages/Home'
 import AllProducts from '../../pages/AllProducts'
 import Yarn from '../../pages/Yarn'
-import CraftTools from '../../pages/CraftTools'
+import CraftTools from '../../pages/Tools'
 import Books from '../../pages/Books'
 import InduvidualProduct from '../../pages/InduvidualProduct'
 import Login from '../../pages/Login'
@@ -29,14 +29,11 @@ function App() {
         <Route path='/' element={<Home />} />
         <Route path='/allproducts' element={<AllProducts />} />
         <Route path='/yarn' element={<Yarn />} />
-        <Route path='/crafttools' element={<CraftTools />} />
+        <Route path='/tools' element={<CraftTools />} />
         <Route path='/books' element={<Books />} />
         <Route path='/cart' element={<Cart />} />
         <Route path='/login' element={<Login />} />
-        <Route path='/product/:id' element={<InduvidualProduct />} />
-        {/* <Route path='/books/:id' element={<InduvidualBook />} />
-        <Route path='/yarn/:id' element={<InduvidualYarn />} />
-        <Route path='/crafttools/:id' element={<InduvidualTool />} /> */}
+        <Route path='/:category/:id' element={<InduvidualProduct />} />
       </Routes>
     </div>
   );
