+++ src/components/NavBar/NavBar.css
@@ -1,4 +1,5 @@
 nav {
+  font-family: 'Bobby Jones Soft';
   display: flex;
   align-items: center;
   justify-content: space-between;
@@ -16,6 +17,7 @@ nav {
 }
 
 .discount {
+  font-family: 'Bobby Jones Soft';
   background-color: #4e2375;
   color: white;
   font-size: 20px;
