+++ src/pages/Books.js
@@ -19,7 +19,8 @@ function getSpecific() {
 
   return (
     <section>
-      <h1>Books</h1>
+      <h1 className='page-title'>Books</h1>
+      <p className='page-description'>Discover a world of knowledge and imagination in our diverse collection of crafting books. Whether you're an avid reader, a curious learner, or searching for the perfect crocheting/knitting project, our carefully curated selection offers something for everyone!</p>
       <div>
         {books.length && <SingleProduct products={books} />}
       </div>
