+++ src/pages/pages.css
@@ -1,4 +1,5 @@
 .home-greeting {
+  font-family: 'Bobby Jones Soft';
   font-size: 45px;
   margin-top: 20px;
   margin-bottom: 0px;
@@ -10,27 +11,57 @@
 }
 
 .ideas-title {
+  font-size: 25px;
   color: black;
   margin-top: 0px
 }
 
 .home-ideas {
+  font-family: 'Bobby Jones Condensed';
   height: contain;
   display: flex;
   flex-wrap: wrap;
   justify-content: center;
 }
 
+.book-name {
+  font-size: 20px;
+  margin: 10px;
+}
+
+.book-price {
+  font-size: 25px;
+  margin: 20px;
+}
+
+.page-description {
+  font-family: 'Bobby Jones Condensed';
+  font-size: 20px;
+  width: 90%;
+  text-align: left;
+  margin: auto;
+  text-align: center;
+}
+
+.page-title {
+  font-family: 'Bobby Jones Condensed';
+  font-size: 32px;
+  text-align: center;
+  margin-bottom: 25px;
+}
+
 .book-container {
-  display: flex;
+  display: grid;
   justify-content: center;
-  align-items: center;
-  flex-wrap: wrap;
+  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+  grid-gap: 32px;
+  grid-auto-flow: dense;  
+  grid-column-gap: 5px;
+  grid-row-gap: 5px;
 }
 
 .book-image {
-  width: 250px;
-  height: 300px;
+  width: 100%;
 }
 
 .yarn-image {
@@ -66,7 +97,7 @@
   width: 200px;
   cursor: pointer;
   font-size: 25px;
-  background-color:#ab8fca;
+  background-color: #ab8fca;
   color: white;
   border: 2px black solid;
   border-radius: 20px;
@@ -91,8 +122,12 @@
 
 .small-book-container {
   margin: 20px;
-  width: 25%;
-  flex-wrap: wrap;
+  padding: 10px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  box-shadow: 3px 3px 3px 3px rgb(213, 209, 209);
+  font-family: 'Bobby Jones Condensed';
   text-decoration: none;
   color: black;
 }
